[Refactoring] #108 Move select configuration logic outside MFSelectConfigurationsAction file
